Added ClearRegistry() to Reset. Cleaned up some files.
